Move PagedScoreSheet to separate file
